requirements.txt
torchvision==0.4.0
pandas==0.25.2
numpy==1.17.2
torch==1.2.0
matplotlib==3.1.1
tqdm==4.41.1
Pillow==7.2.0
scikit_learn==0.23.1
